comit

+++ db/migrate/20200920090715_create_cars.rb
@@ -3,7 +3,7 @@ class CreateCars < ActiveRecord::Migration[6.0]
     create_table :cars do |t|
       t.string :car_name, null: false
       t.string :image
-      t.integer :first_range, null: false
+      t.float :first_range, null: false
       t.references :user, foreign_key: true
       t.timestamps
     end
